Move utility functions to util
