+++ src/util.js
@@ -1,22 +1,40 @@
 import {expandUrl} from 'ol/tileurlfunction.js';
 import {normalizeSourceUrl, normalizeStyleUrl} from './mapbox.js';
+import {styleFunctionArgs, stylefunction} from './stylefunction.js';
 
-let styleId = 0;
+/** @typedef {import("ol").Map} Map */
+/** @typedef {import("ol/layer").Layer} Layer */
+/** @typedef {import("ol/layer").Group} LayerGroup */
+/** @typedef {import("ol/layer").Vector} VectorLayer */
+/** @typedef {import("ol/layer").VectorTile} VectorTileLayer */
+/** @typedef {import("ol/source").Source} Source */
+
+/**
+ * @typedef {Object} FeatureIdentifier
+ * @property {string|number} id The feature id.
+ * @property {string} source The source id.
+ */
+
 const functionCacheByStyleId = {};
 const filterCacheByStyleId = {};
+
+let styleId = 0;
+export function getStyleId(glStyle) {
+  if (!glStyle.id) {
+    glStyle.id = styleId++;
+  }
+  return glStyle.id;
+}
 
 /**
  * @param {Object} glStyle Mapboox style object.
  * @return {Object} Function cache.
  */
 export function getFunctionCache(glStyle) {
-  if (!glStyle.id) {
-    glStyle.id = styleId++;
-  }
   let functionCache = functionCacheByStyleId[glStyle.id];
   if (!functionCache) {
     functionCache = {};
-    functionCacheByStyleId[glStyle.id] = functionCache;
+    functionCacheByStyleId[getStyleId(glStyle)] = functionCache;
   }
   return functionCache;
 }
@@ -32,13 +50,10 @@ export function clearFunctionCache() {
  * @return {Object} Filter cache.
  */
 export function getFilterCache(glStyle) {
-  if (!glStyle.id) {
-    glStyle.id = styleId++;
-  }
   let filterCache = filterCacheByStyleId[glStyle.id];
   if (!filterCache) {
     filterCache = {};
-    filterCacheByStyleId[glStyle.id] = filterCache;
+    filterCacheByStyleId[getStyleId(glStyle)] = filterCache;
   }
   return filterCache;
 }
@@ -276,6 +291,218 @@ export function drawIconHalo(
   }
   imageContext.fill();
   return imageCanvas;
+}
+
+/**
+ * Get the OpenLayers layer instance that contains the provided Mapbox Style
+ * `layer`. Note that multiple Mapbox Style layers are combined in a single
+ * OpenLayers layer instance when they use the same Mapbox Style `source`.
+ * @param {Map|LayerGroup} map OpenLayers Map or LayerGroup.
+ * @param {string} layerId Mapbox Style layer id.
+ * @return {Layer} OpenLayers layer instance.
+ */
+export function getLayer(map, layerId) {
+  const layers = map.getLayers().getArray();
+  for (let i = 0, ii = layers.length; i < ii; ++i) {
+    const mapboxLayers = layers[i].get('mapbox-layers');
+    if (mapboxLayers && mapboxLayers.indexOf(layerId) !== -1) {
+      return /** @type {Layer} */ (layers[i]);
+    }
+  }
+}
+
+/**
+ * Get the OpenLayers layer instances for the provided Mapbox Style `source`.
+ * @param {Map|LayerGroup} map OpenLayers Map or LayerGroup.
+ * @param {string} sourceId Mapbox Style source id.
+ * @return {Array<Layer>} OpenLayers layer instances.
+ */
+export function getLayers(map, sourceId) {
+  const result = [];
+  const layers = map.getLayers().getArray();
+  for (let i = 0, ii = layers.length; i < ii; ++i) {
+    if (layers[i].get('mapbox-source') === sourceId) {
+      result.push(/** @type {Layer} */ (layers[i]));
+    }
+  }
+  return result;
+}
+
+/**
+ * Get the OpenLayers source instance for the provided Mapbox Style `source`.
+ * @param {Map|LayerGroup} map OpenLayers Map or LayerGroup.
+ * @param {string} sourceId Mapbox Style source id.
+ * @return {Source} OpenLayers source instance.
+ */
+export function getSource(map, sourceId) {
+  const layers = map.getLayers().getArray();
+  for (let i = 0, ii = layers.length; i < ii; ++i) {
+    const source = /** @type {Layer} */ (layers[i]).getSource();
+    if (layers[i].get('mapbox-source') === sourceId) {
+      return source;
+    }
+  }
+}
+
+/**
+ * Sets or removes a feature state. The feature state is taken into account for styling,
+ * just like the feature's properties, and can be used e.g. to conditionally render selected
+ * features differently.
+ *
+ * The feature state will be stored on the OpenLayers layer matching the feature identifier, in the
+ * `mapbox-featurestate` property.
+ * @param {Map|VectorLayer|VectorTileLayer} mapOrLayer OpenLayers Map or layer to set the feature
+ * state on.
+ * @param {FeatureIdentifier} feature Feature identifier.
+ * @param {Object|null} state Feature state. Set to `null` to remove the feature state.
+ */
+export function setFeatureState(mapOrLayer, feature, state) {
+  const layers =
+    'getLayers' in mapOrLayer
+      ? getLayers(mapOrLayer, feature.source)
+      : [mapOrLayer];
+  for (let i = 0, ii = layers.length; i < ii; ++i) {
+    const featureState = layers[i].get('mapbox-featurestate');
+    if (featureState) {
+      if (state) {
+        featureState[feature.id] = state;
+      } else {
+        delete featureState[feature.id];
+      }
+      layers[i].changed();
+    } else {
+      throw new Error(`Map or layer for source "${feature.source}" not found.`);
+    }
+  }
+}
+
+/**
+ * Sets or removes a feature state. The feature state is taken into account for styling,
+ * just like the feature's properties, and can be used e.g. to conditionally render selected
+ * features differently.
+ * @param {Map|VectorLayer|VectorTileLayer} mapOrLayer Map or layer to set the feature state on.
+ * @param {FeatureIdentifier} feature Feature identifier.
+ * @return {Object|null} Feature state or `null` when no feature state is set for the given
+ * feature identifier.
+ */
+export function getFeatureState(mapOrLayer, feature) {
+  const layers =
+    'getLayers' in mapOrLayer
+      ? getLayers(mapOrLayer, feature.source)
+      : [mapOrLayer];
+  for (let i = 0, ii = layers.length; i < ii; ++i) {
+    const featureState = layers[i].get('mapbox-featurestate');
+    if (featureState && featureState[feature.id]) {
+      return featureState[feature.id];
+    }
+  }
+  return null;
+}
+
+/**
+ * Get the Mapbox Layer object for the provided `layerId`.
+ * @param {Map|LayerGroup} mapOrGroup Map or LayerGroup.
+ * @param {string} layerId Mapbox Layer id.
+ * @return {Object} Mapbox Layer object.
+ */
+export function getMapboxLayer(mapOrGroup, layerId) {
+  const style = mapOrGroup.get('mapbox-style');
+  const layerStyle = style.layers.find(function (layer) {
+    return layer.id === layerId;
+  });
+  return layerStyle;
+}
+
+/**
+ * Add a new Mapbox Layer object to the style. The map will be re-rendered.
+ * @param {Map|LayerGroup} mapOrGroup The Map or LayerGroup `apply` was called on.
+ * @param {Object} mapboxLayer Mapbox Layer object.
+ * @param {string} [beforeLayerId] Optional id of the Mapbox Layer before the new layer that will be added.
+ */
+export function addMapboxLayer(mapOrGroup, mapboxLayer, beforeLayerId) {
+  const mapboxLayers = mapOrGroup.get('mapbox-style').layers;
+  let index;
+  if (beforeLayerId !== undefined) {
+    const beforeLayer = getMapboxLayer(mapOrGroup, beforeLayerId);
+    if (beforeLayer === undefined) {
+      throw new Error(`Layer with id "${beforeLayerId}" not found.`);
+    }
+    index = mapboxLayers.indexOf(beforeLayer);
+  } else {
+    index = mapboxLayers.length;
+  }
+  if (index === 0) {
+    throw new Error('Cannot add layer before first layer.');
+  }
+  if (mapboxLayers[index - 1].source !== mapboxLayer.source) {
+    throw new Error('Added layer and layer before must use the same source.');
+  }
+  if (mapboxLayers.some((layer) => layer.id === mapboxLayer.id)) {
+    throw new Error(`Layer with id "${mapboxLayer.id}" already exists.`);
+  }
+  mapboxLayers.splice(index, 0, mapboxLayer);
+  stylefunction.apply(
+    undefined,
+    styleFunctionArgs[getStyleId(mapOrGroup.get('mapbox-style'))]
+  );
+}
+
+/**
+ * Update a Mapbox Layer object in the style. The map will be re-rendered with the new style.
+ * @param {Map|LayerGroup} mapOrGroup The Map or LayerGroup `apply` was called on.
+ * @param {Object} mapboxLayer Updated Mapbox Layer object.
+ */
+export function updateMapboxLayer(mapOrGroup, mapboxLayer) {
+  const glStyle = mapOrGroup.get('mapbox-style');
+  const mapboxLayers = glStyle.layers;
+  const index = mapboxLayers.findIndex(function (layer) {
+    return layer.id === mapboxLayer.id;
+  });
+  if (index === -1) {
+    throw new Error(`Layer with id "${mapboxLayer.id}" not found.`);
+  }
+  const oldLayer = mapboxLayers[index];
+  if (oldLayer.source !== mapboxLayer.source) {
+    throw new Error(
+      'Updated layer and previous version must use the same source.'
+    );
+  }
+  delete getFunctionCache(glStyle)[mapboxLayer.id];
+  delete getFilterCache(glStyle)[mapboxLayer.id];
+  stylefunction.apply(
+    undefined,
+    styleFunctionArgs[getStyleId(mapOrGroup.get('mapbox-style'))]
+  );
+}
+
+/**
+ * Remove a Mapbox Layer object from the style. The map will be re-rendered.
+ * @param {Map|LayerGroup} mapOrGroup The Map or LayerGroup `apply` was called on.
+ * @param {string|Object} mapboxLayerIdOrLayer Mapbox Layer id or Mapbox Layer object.
+ */
+export function removeMapboxLayer(mapOrGroup, mapboxLayerIdOrLayer) {
+  const mapboxLayerId =
+    typeof mapboxLayerIdOrLayer === 'string'
+      ? mapboxLayerIdOrLayer
+      : mapboxLayerIdOrLayer.id;
+  const layer = getLayer(mapOrGroup, mapboxLayerId);
+  /** @type {Array<Object>} */
+  const layerMapboxLayers = layer.get('mapbox-layers');
+  if (layerMapboxLayers.length === 1) {
+    throw new Error(
+      'Cannot remove last Mapbox layer from an OpenLayers layer.'
+    );
+  }
+  layerMapboxLayers.splice(layerMapboxLayers.indexOf(mapboxLayerId), 1);
+  const layers = mapOrGroup.get('mapbox-style').layers;
+  layers.splice(
+    layers.findIndex((layer) => layer.id === mapboxLayerId),
+    1
+  );
+  stylefunction.apply(
+    undefined,
+    styleFunctionArgs[getStyleId(mapOrGroup.get('mapbox-style'))]
+  );
 }
 
 /**
